+++ .env.sample
@@ -143,3 +143,8 @@ DIAGORIENTE_API_CLIENT_SECRET=if_fake_client_secret
 
 CACHE=REDIS
 REDIS_URL=redis://:redis_password@localhost:6379
+
+CRISP_GATEWAY=IN_MEMORY
+CRISP_ID=crisp-id
+CRISP_KEY=crisp-key
+CRISP_WEBSITE_ID=some-website-id
